test(add-error): realocate test
