+++ src/Features/SupportValidation/UnitTest.php
@@ -769,6 +769,31 @@ class UnitTest extends \Tests\TestCase
             ->assertHasNoErrors('customCollection.0.amount')
             ;
     }
+
+    /** @test */
+    public function adding_validation_error_inside_mount_method()
+    {
+        Livewire::test(AddErrorInMount::class)
+            ->call('addErrors')
+            ->assertSee('first error')
+            ->assertSee('second error')
+            ->assertSee('third error')
+            ->assertHasErrors(['first', 'second', 'third'])
+            ->call('addFilterErrors')
+            ->assertSee('first error')
+            ->assertSee('second error')
+            ->assertHasErrors(['first', 'second']);
+
+        Livewire::test(AddErrorInMount::class)
+            ->assertSee('first error')
+            ->assertSee('second error')
+            ->assertSee('third error')
+            ->assertHasErrors(['first', 'second', 'third'])
+            ->call('addFilterErrors')
+            ->assertSee('first error')
+            ->assertSee('second error')
+            ->assertHasErrors(['first', 'second']);
+    }
 }
 
 class ComponentWithRulesProperty extends Component
@@ -1276,5 +1301,33 @@ class CustomWireableValidationDTO implements Wireable
         return new static(
             $value['amount']
         );
+    }
+}
+
+class AddErrorInMount extends Component
+{
+    public function mount()
+    {
+        $this->addErrors();
+    }
+
+    public function addErrors(): void
+    {
+        $this->addError('first', 'first error');
+        $this->addError('second', 'second error');
+        $this->addError('third', 'third error');
+    }
+
+    public function addFilterErrors(): void
+    {
+        $this->resetErrorBag();
+
+        $this->addError('first', 'first error');
+        $this->addError('second', 'second error');
+    }
+
+    public function render()
+    {
+        return view('show-errors');
     }
 }
